+++ src/FuseDigital.QuickSetup.Infrastructure/Yaml/IYamlContext.cs
