+++ src/FuseDigital.QuickSetup.Infrastructure/Yaml/IYamlRepository.cs
