+++ src/FuseDigital.QuickSetup.Infrastructure/Yaml/YamlEntityExtensions.cs
