+++ test/FuseDigital.QuickSetup.Infrastructure.Tests/QuickSetupInfrastructureTestModule.cs
@@ -4,7 +4,7 @@ namespace FuseDigital.QuickSetup;
 
 [DependsOn(
     typeof(QuickSetupTestBaseModule),
-    typeof(QuickSetupYamlModule))]
-public class QuickSetupYamlTestModule : AbpModule
+    typeof(QuickSetupInfrastructureModule))]
+public class QuickSetupInfrastructureTestModule : AbpModule
 {
 }
