+++ test/FuseDigital.QuickSetup.Infrastructure.Tests/Yaml/sample-list.yml
